refactor: 할 일 추가 버튼 위치 변경
